Add code for tracking systematic uncertainty (#123)
* Add script to create input histos for tracking syst unc

* Add script for tracking uncertainty propagation

* Fix typo in comment

* Add util function for application of histo values to pandas

* Forse pandas type to allow query with negative values and bitmap operations

* Restore normalisation in feature distributions

* Revert wrong modification

* Improve drawing style and add legends

* Remove unneeded enumerated

+++ filterdata/FilterTree4ML.py
@@ -59,7 +59,7 @@ print('Applying selections')
 dataFramePtCut = dataFrame.query(f'pt_cand > {PtMin} & pt_cand < {PtMax}')
 del dataFrame
 if preSelections:
-    dataFramePtCutSel = dataFramePtCut.query(preSelections)
+    dataFramePtCutSel = dataFramePtCut.astype(float).query(preSelections)
     del dataFramePtCut
 else:
     dataFramePtCutSel = dataFramePtCut
